requirements.txt
PyQt5
sympy
numpy
tabulate
antlr4-python3-runtime 